feat: 2FA L1 integration (#726)
## Summary

Use signatures collected through 2FA Batch Verification for submission
through `commitBatchesMultisig`. Validator set and threshold are also
received from L1.

+++ integration-tests/src/lib.rs
@@ -382,7 +382,7 @@ impl Tester {
 pub struct TesterBuilder {
     enable_prover: bool,
     block_time: Option<Duration>,
-    batch_verification_threshold: Option<usize>,
+    batch_verification_threshold: Option<u64>,
 }
 
 impl TesterBuilder {
@@ -397,7 +397,7 @@ impl TesterBuilder {
         self
     }
 
-    pub fn batch_verification(mut self, threshold: usize) -> Self {
+    pub fn batch_verification(mut self, threshold: u64) -> Self {
         self.batch_verification_threshold = Some(threshold);
         self
     }
